% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean}
\alias{mean}
\title{Mean}
\usage{
mean(f)
}
\arguments{
\item{f(x)}{A probability density function}
}
\description{
This function calculates the mean of a continuous random variable
}
\examples{
mean(function (x) {
     fun <- 3 * x ^ 2
     fun[x < 0] = 0
     fun[x > 1] = 0
     return ( fun )
})
}
\keyword{mean}
