% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.R
\name{P}
\alias{P}
\title{Probability}
\usage{
P(f, p)
}
\arguments{
\item{p}{A string, indicating a (conditional) probability.}

\item{f(x)}{A probability density function}
}
\description{
This function calculates (conditional) probabilities of a continuous random variable.
}
\examples{
f <- function (x) {
     fun <- 3 * x ^ 2
     fun[x < 0] = 0
     fun[x > 1] = 0
     return ( fun )
}

P(test, "X < 1 | X < 3")  # good
P(test, "X < 2")          # good
P(test, "X < 1 | X > 1")  # good
P(test, "X < 1 | X Z")    # wrong
P(test, "X < 1 | Y < 2")  # wrong
P(test, "X < -1 & X > 1") # wrong
}
\keyword{p}
\keyword{prob}
\keyword{probability}
