% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\name{var}
\alias{var}
\title{Variance}
\usage{
var(f)
}
\arguments{
\item{f(x)}{A probability density function}
}
\description{
This function calculates the variance of a continuous random variable.
}
\examples{
var(function (x) {
     fun <- 3 * x ^ 2
     fun[x < 0] = 0
     fun[x > 1] = 0
     return ( fun )
})
}
\keyword{var}
\keyword{variance}
